feat: 历史记录和弹幕相关接口

+++ server/internal/domain/dto/danmaku.go
@@ -3,8 +3,8 @@ package dto
 type DanmakuReq struct {
 	Vid   uint
 	Part  uint
-	Time  uint //时间
-	Type  int  //类型0滚动;1顶部;2底部
+	Time  float32 //时间
+	Type  int     //类型0滚动;1顶部;2底部
 	Color string
 	Text  string
 }
